third_party/blink/web_tests/http/tests/websocket/url-with-query-expected.txt

Make sure handshake with URL with query components success.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

url=ws://127.0.0.1:8880/echo-path?q=v
WebSocket is open
received:/echo-path?q=v
WebSocket is closed
PASS handshake_success is true
PASS path is expected_path
PASS successfullyParsed is true

TEST COMPLETE
